auth_service/requirements.txt
Flask==2.2.5
Flask-Limiter==3.3.1
redis==4.5.5
python-dotenv==1.0.0
psycopg2-binary==2.9.6
Flask-SQLAlchemy==3.0.2
Flask-JWT-Extended==4.4.4